requirements.txt
Flask==1.1.2
awscli==1.18.23
boto3==1.12.23
boto==2.49.0
pycryptodome==3.9.7
pymongo==3.11.0
pyjwt==2.0.0
requests==2.25.1
cryptography==3.3.1
